Remove fatal error in creating wrapper for extern functions with array args
I added a fatal error to check if a code path was still reached. It is still
reached, but can be safely handled by opposite branch of the conditional.

Also, fixed a test problem with a new test.

+++ test/extern/diten/testVoidExternFns.h
@@ -2,9 +2,9 @@
 #include <stdint.h>
 
 static void voidNoArray(int64_t n) {
-  printf("%lld\n", n);
+  printf("%d\n", (int)n);
 }
 
 static void voidWithArray(int64_t* A, int64_t n) {
-  printf("%lld %lld %lld\n", A[0], A[1], n);
+  printf("%d %d %d\n", (int)A[0], (int)A[1], (int)n);
 }
